+++ tool/cmd_latex.ml
@@ -112,10 +112,10 @@ Buffer.add_string (if files <> [] then "\n" else "")
 let latex
     files quiet accumulate_defs strict heading_auto_ids backend_blocks
     lift_headings docu title author inline_preambles keep_built_in_preambles
-    first_heading
+    first_heading_level
   =
   let resolver = Label_resolver.v ~quiet in
-  let r = Cmarkit_latex.renderer ~backend_blocks ~first_heading () in
+  let r = Cmarkit_latex.renderer ~backend_blocks ~first_heading_level () in
   let parse ~extract_title ~file ~defs md =
     let doc =
       Cmarkit.Doc.of_string ~resolver ~defs ~heading_auto_ids ~file ~strict md
@@ -179,19 +179,18 @@ let lift_headings =
   in
   Arg.(value & flag & info ["l"; "lift-headings"] ~doc)
 
-let first_heading =
-  let doc = "The first heading kind to use."
+let first_level_heading =
+  let cmd_enum =
+    [ "part", Cmarkit_latex.Part; "chapter", Chapter;
+      "section", Section; "subsection", Subsection ]
   in
-  let parse s =
-    match Cmarkit_latex.heading_of_string s with
-    | None -> Error (`Msg (Format.sprintf "invalid heading kind: %s" s))
-    | Some h -> Ok h
+  let doc =
+    Printf.sprintf
+      "Use LaTeX heading level $(docv) for the first CommonMark heading level. \
+       $(docv) must be %s." (Arg.doc_alts_enum cmd_enum)
   in
-  let print = Cmarkit_latex.pp_heading in
-  let c = Arg.conv ~docv:"VALUE" (parse, print) in
-  let v = Cmarkit_latex.Section in
-  let i = Arg.info ["f"; "first-heading"] ~doc in
-  Arg.(value & opt c v i)
+  Arg.(value & opt (Arg.enum cmd_enum) Cmarkit_latex.Section &
+       Arg.info ["first-heading-level"] ~doc ~docv:"LEVEL")
 
 let v =
   let doc = "Render CommonMark to LaTeX" in
@@ -208,7 +207,7 @@ let v =
   Term.(const latex $ Cli.files $ Cli.quiet $ Cli.accumulate_defs $ Cli.strict $
         Cli.heading_auto_ids $ backend_blocks $ lift_headings $
         Cli.docu $ Cli.title $ author $ inline_preambles $
-        keep_built_in_preamble $ first_heading)
+        keep_built_in_preamble $ first_level_heading)
 
 (* Built-in LaTeX preamable, defined that way to avoid source clutter *)